Bugfix on houses offset calculation.

+++ src/34_road/MainRoadChunk.cpp
@@ -116,12 +116,11 @@ namespace synthese
 				return startOffset;
 			}
 
-			return
-				startOffset + (endOffset - startOffset) * (
-					(bounds->first < bounds->second) ?
-					((houseNumber - bounds->first) / (bounds->second - bounds->first)) :
-					((bounds->first - houseNumber) / (bounds->first - bounds->second))
-				)
+			return startOffset + (endOffset - startOffset) * (
+					   (bounds->first < bounds->second) ?
+					   ((double)(houseNumber - bounds->first) / (double)(bounds->second - bounds->first)) :
+					   ((double)(bounds->first - houseNumber) / (double)(bounds->first - bounds->second))
+				   )
 			;
 		}
